Replace combined base version with async
async-beta
