+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.5
 
 import PackageDescription
 
@@ -10,146 +10,57 @@ let package = Package(
     targets: Target.targets
 )
 
-#if swift(>=5.5)
-    extension Product {
-        static let products: [Product] = [
-            .library(
-                name: "NetworkService",
-                targets: ["NetworkService"]
-            ),
-            .library(
-                name: "NetworkServiceTestHelper",
-                targets: ["NetworkServiceTestHelper"]
-            ),
-            .library(
-                name: "NetworkServiceAsyncBeta",
-                targets: ["NetworkServiceAsyncBeta"]
-            ),
-            .library(
-                name: "NetworkServiceTestHelperAsyncBeta",
-                targets: ["NetworkServiceTestHelperAsyncBeta"]
-            ),
-        ]
-    }
+extension Product {
+    static let products: [Product] = [
+        .library(
+            name: "NetworkService",
+            targets: ["NetworkService"]
+        ),
+        .library(
+            name: "NetworkServiceTestHelper",
+            targets: ["NetworkServiceTestHelper"]
+        ),
+    ]
+}
 
-    extension Target {
-        static let targets: [Target] = [
-            .target(
-                name: "NetworkService",
-                dependencies: []
-            ),
-            .testTarget(
-                name: "NetworkServiceTests",
-                dependencies: [
-                    "NetworkService",
-                    .product(name: "OHHTTPStubs", package: "OHHTTPStubs"),
-                    .product(name: "OHHTTPStubsSwift", package: "OHHTTPStubs"),
-                ]
-            ),
-            .target(name: "NetworkServiceAsyncBeta"),
-            .testTarget(
-                name: "NetworkServiceAsyncBetaTests",
-                dependencies: [
-                    "NetworkServiceAsyncBeta",
-                    .product(name: "OHHTTPStubs", package: "OHHTTPStubs"),
-                    .product(name: "OHHTTPStubsSwift", package: "OHHTTPStubs"),
-                ]
-            ),
-            .target(
-                name: "NetworkServiceTestHelper",
-                dependencies: [
-                    "NetworkService",
-                    .product(name: "CombineSchedulers", package: "combine-schedulers"),
-                ]
-            ),
-            .testTarget(
-                name: "NetworkServiceTestHelperTests",
-                dependencies: [
-                    "NetworkServiceTestHelper",
-                    .product(name: "CombineSchedulers", package: "combine-schedulers"),
-                ]
-            ),
-            .target(
-                name: "NetworkServiceTestHelperAsyncBeta",
-                dependencies: [
-                    "NetworkServiceAsyncBeta",
-                    .product(name: "CombineSchedulers", package: "combine-schedulers"),
-                ]
-            ),
-            .testTarget(
-                name: "NetworkServiceTestHelperAsyncBetaTests",
-                dependencies: [
-                    "NetworkServiceTestHelperAsyncBeta",
-                    .product(name: "CombineSchedulers", package: "combine-schedulers"),
-                ]
-            ),
-        ]
-    }
+extension Target {
+    static let targets: [Target] = [
+        .target(
+            name: "NetworkService",
+            dependencies: []
+        ),
+        .testTarget(
+            name: "NetworkServiceTests",
+            dependencies: [
+                "NetworkService",
+                .product(name: "OHHTTPStubs", package: "OHHTTPStubs"),
+                .product(name: "OHHTTPStubsSwift", package: "OHHTTPStubs"),
+            ]
+        ),
+        .target(
+            name: "NetworkServiceTestHelper",
+            dependencies: [
+                "NetworkService",
+                .product(name: "CombineSchedulers", package: "combine-schedulers"),
+            ]
+        ),
+        .testTarget(
+            name: "NetworkServiceTestHelperTests",
+            dependencies: [
+                "NetworkServiceTestHelper",
+                .product(name: "CombineSchedulers", package: "combine-schedulers"),
+            ]
+        ),
+    ]
+}
 
-    extension Package.Dependency {
-        static let dependencies: [Package.Dependency] = [
-            .package(url: "https://github.com/AliSoftware/OHHTTPStubs.git", from: "9.1.0"),
-            .package(
-                name: "combine-schedulers",
-                url: "https://github.com/pointfreeco/combine-schedulers.git",
-                .upToNextMajor(from: "0.6.0")
-            ),
-        ]
-    }
-
-#else
-    extension Product {
-        static let products: [Product] = [
-            .library(
-                name: "NetworkService",
-                targets: ["NetworkService"]
-            ),
-            .library(
-                name: "NetworkServiceTestHelper",
-                targets: ["NetworkServiceTestHelper"]
-            ),
-        ]
-    }
-
-    extension Target {
-        static let targets: [Target] = [
-            .target(
-                name: "NetworkService",
-                dependencies: []
-            ),
-            .testTarget(
-                name: "NetworkServiceTests",
-                dependencies: [
-                    "NetworkService",
-                    .product(name: "OHHTTPStubs", package: "OHHTTPStubs"),
-                    .product(name: "OHHTTPStubsSwift", package: "OHHTTPStubs"),
-                ]
-            ),
-            .target(
-                name: "NetworkServiceTestHelper",
-                dependencies: [
-                    "NetworkService",
-                    .product(name: "CombineSchedulers", package: "combine-schedulers"),
-                ]
-            ),
-            .testTarget(
-                name: "NetworkServiceTestHelperTests",
-                dependencies: [
-                    "NetworkServiceTestHelper",
-                    .product(name: "CombineSchedulers", package: "combine-schedulers"),
-                ]
-            ),
-        ]
-    }
-
-    extension Package.Dependency {
-        static let dependencies: [Package.Dependency] = [
-            .package(url: "https://github.com/AliSoftware/OHHTTPStubs.git", from: "9.1.0"),
-            .package(
-                name: "combine-schedulers",
-                url: "https://github.com/pointfreeco/combine-schedulers.git",
-                .upToNextMinor(from: "0.5.3")
-            ),
-        ]
-    }
-#endif
+extension Package.Dependency {
+    static let dependencies: [Package.Dependency] = [
+        .package(url: "https://github.com/AliSoftware/OHHTTPStubs.git", from: "9.1.0"),
+        .package(
+            name: "combine-schedulers",
+            url: "https://github.com/pointfreeco/combine-schedulers.git",
+            .upToNextMajor(from: "0.6.0")
+        ),
+    ]
+}
