+++ Sources/NetworkService/NetworkServiceClient+Put.swift
@@ -21,12 +21,9 @@ extension NetworkServiceClient {
         _ body: Data,
         to url: URL,
         headers: [HTTPHeader] = []
-    ) -> AnyPublisher<Data, Failure> {
-        var request = URLRequest(url: url)
-        request.httpBody = body
-        request.method = .PUT
-        headers.forEach { request.addValue($0) }
-        return start(request)
+    ) async -> Result<Data, Failure> {
+        let request = URLRequest.build(url: url, body: body, headers: headers, method: .PUT)
+        return await start(request)
     }
 
     /// - Parameters:
@@ -40,17 +37,18 @@ extension NetworkServiceClient {
         to url: URL,
         headers: [HTTPHeader],
         encoder: Encoder
-    ) -> AnyPublisher<Data, Failure>
+    ) async -> Result<Data, Failure>
         where RequestBody: Encodable,
         Encoder: TopLevelEncoder,
         Encoder.Output == Data
     {
         do {
             let body = try encoder.encode(body)
-            return put(body, to: url, headers: headers)
+            return await put(body, to: url, headers: headers)
+        } catch let urlError as URLError {
+            return .failure(Failure.urlError(urlError))
         } catch {
-            return Fail(error: Failure.cocoa(error as NSError))
-                .eraseToAnyPublisher()
+            return .failure(Failure.unknown(error as NSError))
         }
     }
 
@@ -63,15 +61,16 @@ extension NetworkServiceClient {
         _ body: RequestBody,
         to url: URL,
         headers: [HTTPHeader]
-    ) -> AnyPublisher<Data, Failure>
+    ) async -> Result<Data, Failure>
         where RequestBody: TopLevelEncodable
     {
         do {
             let body = try RequestBody.encoder.encode(body)
-            return put(body, to: url, headers: headers)
+            return await put(body, to: url, headers: headers)
+        } catch let urlError as URLError {
+            return .failure(Failure.urlError(urlError))
         } catch {
-            return Fail(error: Failure.cocoa(error as NSError))
-                .eraseToAnyPublisher()
+            return .failure(Failure.unknown(error as NSError))
         }
     }
 
@@ -87,13 +86,12 @@ extension NetworkServiceClient {
         to url: URL,
         headers: [HTTPHeader] = [],
         decoder: Decoder
-    ) -> AnyPublisher<ResponseBody, Failure>
+    ) async -> Result<ResponseBody, Failure>
         where ResponseBody: Decodable, Decoder: TopLevelDecoder, Decoder.Input == Data
     {
-        put(body, to: url, headers: headers)
+        await put(body, to: url, headers: headers)
             .decode(with: decoder)
             .mapToNetworkError()
-            .eraseToAnyPublisher()
     }
 
     /// - Parameters:
@@ -105,10 +103,10 @@ extension NetworkServiceClient {
         _ body: Data,
         to url: URL,
         headers: [HTTPHeader] = []
-    ) -> AnyPublisher<ResponseBody, Failure>
+    ) async -> Result<ResponseBody, Failure>
         where ResponseBody: TopLevelDecodable
     {
-        put(body, to: url, headers: headers, decoder: ResponseBody.decoder)
+        await put(body, to: url, headers: headers, decoder: ResponseBody.decoder)
     }
 
     /// Send a put request to a `URL`
@@ -125,7 +123,7 @@ extension NetworkServiceClient {
         headers: [HTTPHeader] = [],
         encoder: Encoder,
         decoder: Decoder
-    ) -> AnyPublisher<ResponseBody, Failure>
+    ) async -> Result<ResponseBody, Failure>
         where RequestBody: Encodable,
         ResponseBody: Decodable,
         Encoder: TopLevelEncoder,
@@ -135,10 +133,11 @@ extension NetworkServiceClient {
     {
         do {
             let body = try encoder.encode(body)
-            return put(body, to: url, headers: headers, decoder: decoder)
+            return await put(body, to: url, headers: headers, decoder: decoder)
+        } catch let urlError as URLError {
+            return .failure(Failure.urlError(urlError))
         } catch {
-            return Fail(error: Failure.cocoa(error as NSError))
-                .eraseToAnyPublisher()
+            return .failure(Failure.unknown(error as NSError))
         }
     }
 
@@ -152,10 +151,14 @@ extension NetworkServiceClient {
         _ body: RequestBody,
         to url: URL,
         headers: [HTTPHeader] = []
-    ) -> AnyPublisher<ResponseBody, Failure>
+    ) async -> Result<ResponseBody, Failure>
         where RequestBody: TopLevelEncodable,
         ResponseBody: TopLevelDecodable
     {
-        put(body, to: url, headers: headers, encoder: RequestBody.encoder, decoder: ResponseBody.decoder)
+        await put(body, to: url, headers: headers, encoder: RequestBody.encoder, decoder: ResponseBody.decoder)
     }
 }
+
+#if canImport(Combine)
+    extension NetworkServiceClient {}
+#endif
