+++ Sources/NetworkService/Result+NetworkService.swift
