+++ Sources/NetworkServiceTestHelper/CodableOutput.swift
@@ -7,7 +7,7 @@
 // LICENSE file in the root directory of this source tree.
 
 import Foundation
-import NetworkServiceAsyncBeta
+import NetworkService
 
 #if canImport(Combine)
     import Combine
