+++ Sources/NetworkServiceTestHelper/FailureOutput.swift
@@ -7,7 +7,7 @@
 // LICENSE file in the root directory of this source tree.
 
 import Foundation
-import NetworkServiceAsyncBeta
+import NetworkService
 
 /// Wraps a provided error for use as output
 public struct FailureOutput<T>: MockOutput where T: Error {
