+++ Sources/NetworkServiceTestHelper/MockNetworkService.swift
@@ -14,6 +14,7 @@ import NetworkService
 /// Convenience implementation of `NetworkServiceClient` for testing. Supports defining set output values for all
 /// network functions,
 /// repeating values, and delaying values.
+@available(swift 5.5)
 open class MockNetworkService<T: Scheduler>: NetworkServiceClient {
     public var delay: Delay
     public var outputs: [MockOutput]
@@ -48,131 +49,34 @@ open class MockNetworkService<T: Scheduler>: NetworkServiceClient {
     /// Replaces default implementation from protocol. All `NetworkService` functions should eventually end up in this
     /// version of `start`.
     /// Delay and repeat are handled here.
-    public func start(_: URLRequest) -> AnyPublisher<Data, Failure> {
+    public func start(_: URLRequest) async -> Result<Data, Failure> {
         let next: MockOutput
         do {
             next = try queue()
         } catch {
-            return Fail(error: .cocoa(error as NSError))
-                .eraseToAnyPublisher()
+            return .failure(Failure.unknown(error as NSError))
         }
         switch delay {
-        case .infinite, .seconds:
-            return next.output.publisher
-                .delay(
-                    for: T.SchedulerTimeType.Stride.seconds(delay.interval),
-                    scheduler: scheduler
-                )
-                .receive(on: scheduler)
-                .eraseToAnyPublisher()
+        case .infinite:
+            return await Task {
+                try await scheduler.sleep(for: .seconds(.max))
+                return try next.output.get()
+            }
+            .result.mapToNetworkError()
+        case .seconds:
+            return await Task {
+                try await scheduler.sleep(for: .seconds(delay.interval))
+                return try next.output.get()
+            }
+            .result.mapToNetworkError()
         case .none:
             // Setting the delay publisher to zero seconds was buggy.
             // It works better to not add delay for `none`.
-            return next.output.publisher.eraseToAnyPublisher()
+            return next.output
         }
     }
 
     public enum Errors: Error, Equatable {
         case noOutputQueued
-    }
-}
-
-/// Represents the amount of async delay should be added to the mocked network functions. Consider replacing with
-/// `DispatchTimeInterval`.
-/// Although, there is no included case for zero/none.
-public enum Delay {
-    case infinite
-    case seconds(Int)
-    case none
-
-    /// The delay in `DispatchTimeInterval` for use in schedulars
-    var interval: Int {
-        switch self {
-        case .infinite:
-            return .max
-        case let .seconds(seconds):
-            return seconds
-        case .none:
-            return 0
-        }
-    }
-}
-
-/// Wraps a given output value to define how many times it should be repeated.
-public enum RepeatResponse: MockOutput {
-    case `repeat`(MockOutput, count: Int)
-    case repeatInfinite(MockOutput)
-
-    public var output: Result<Data, NetworkService.Failure> {
-        switch self {
-        case .repeat(let output, count: _), let .repeatInfinite(output):
-            return output.output
-        }
-    }
-}
-
-/// Wraps a provided error for use as output
-public struct FailureOutput<T>: MockOutput where T: Error {
-    public var output: Result<Data, NetworkService.Failure> {
-        if let networkFailure = error as? NetworkService.Failure {
-            return .failure(networkFailure)
-        }
-        return .failure(.cocoa(error as NSError))
-    }
-
-    public let error: T
-
-    public init(error: T) {
-        self.error = error
-    }
-}
-
-/// Fundamental wrapper for output values so they can easily be handled by `MockNetworkService`
-public struct CodableOutput<Output: Codable, Encoder: TopLevelEncoder, Decoder: TopLevelDecoder>: MockOutput
-    where Encoder.Output == Data, Decoder.Input == Data
-{
-    public var output: Result<Data, NetworkService.Failure> {
-        Result {
-            try encoder.encode(value)
-        }
-        .mapError { error in
-            .cocoa(error as NSError)
-        }
-    }
-
-    let value: Output
-    let encoder: Encoder
-
-    public init(_ value: Output, encoder: Encoder) {
-        self.value = value
-        self.encoder = encoder
-    }
-}
-
-/// A type erasing protocol for `MockNetworkService`'s output queue. Allows a heterogenous array.
-public protocol MockOutput {
-    var output: Result<Data, NetworkService.Failure> { get }
-}
-
-extension MockOutput where Self: TopLevelEncodable {
-    public var output: Result<Data, NetworkService.Failure> {
-        Result {
-            try Self.encoder.encode(self)
-        }
-        .mapError { error in
-            .cocoa(error as NSError)
-        }
-    }
-}
-
-extension Data: MockOutput {
-    public var output: Result<Data, NetworkService.Failure> {
-        .success(self)
-    }
-}
-
-extension NetworkService.Failure: MockOutput {
-    public var output: Result<Data, Self> {
-        .failure(self)
     }
 }
