--- Tests/NetworkServiceTests/AsyncTestCase.swift
@@ -1,15 +1,0 @@
-// AsyncTestCase.swift
-// NetworkService
-//
-// Copyright © 2023 MFB Technologies, Inc. All rights reserved.
-//
-// This source code is licensed under the MIT license found in the
-// LICENSE file in the root directory of this source tree.
-
-import Combine
-import Foundation
-import XCTest
-
-class AsyncTestCase: XCTestCase {
-    var cancellables = [AnyCancellable]()
-}
